Complete Adding Command Tutorial
Note: The contents of this commit is insignificant as the work for this
tutorial was added directly to the master branch by accident. This
branch just serves as a reminder that the tutorial has in fact been
completed and in the case where possible changes may be made to the
particular branch.